Swish
Added automatic metadata generation. Ugly but describes the operations recognised that produced an output.
Metadata is saved in a file with the same name as the output file appended with "_SwishMetadata.xml".
Each operation generates metadata describing the operation, arguments, and inputs.
Metadata of input items is aggregated and carried forward.